move chart to a dumb component
